MVVM fixed
