+++ app/build.gradle
@@ -58,7 +58,7 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.5.1"
     //Coroutines
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1'
-    //Hilt
+    // Hilt
     implementation "com.google.dagger:hilt-android:2.40.1"
     kapt "com.google.dagger:hilt-android-compiler:2.40.1"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
