+++ app/src/main/res/drawable/ic_launcher_background.xml
@@ -168,3 +168,4 @@
         android:strokeWidth="0.8"
         android:strokeColor="#33FFFFFF" />
 </vector>
+
